examples/python/requirements.txt
librosa
grpcio
protobuf
